Types

Operator Enum

All equations are represented as a tree of operators. Each node in this tree specifies its operator with an integer - which indexes an enum of operators. This enum is defined as follows:

Missing docstring.

Missing docstring for OperatorEnum{A,B,dA,dB}. Check Documenter's build log for details.

Construct this operator specification as follows:

Missing docstring.

Missing docstring for OperatorEnum(;). Check Documenter's build log for details.

Equations

Equations are specified as binary trees with the Node type, defined as follows:

DynamicExpressions.EquationModule.Node — Type

Node{T<:Real}

Node defines a symbolic expression stored in a binary tree. A single Node instance is one "node" of this tree, and has references to its children. By tracing through the children nodes, you can evaluate or print a given expression.

Fields

degree::Int: Degree of the node. 0 for constants, 1 for unary operators, 2 for binary operators.

constant::Bool: Whether the node is a constant.

val::T: Value of the node. If degree==0, and constant==true, this is the value of the constant. It has a type specified by the overall type of the Node (e.g., Float64).

feature::Int (optional): Index of the feature to use in the case of a feature node. Only used if degree==0 and constant==false. Only defined if degree == 0 && constant == false.

op::Int: If degree==1, this is the index of the operator in operators.unaops. If degree==2, this is the index of the operator in operators.binops. In other words, this is an enum of the operators, and is dependent on the specific OperatorEnum object. Only defined if degree >= 1

l::Node{T}: Left child of the node. Only defined if degree >= 1. Same type as the parent node.

r::Node{T}: Right child of the node. Only defined if degree == 2. Same type as the parent node. This is to be passed as the right argument to the binary operator.

There are a variety of constructors for Node objects, including:

DynamicExpressions.EquationModule.Node — Method

Node(; val::Real=nothing, feature::Integer=nothing)

Create a leaf node: either a constant, or a variable.

Arguments:

val::Real, if you are specifying a constant, pass the value of the constant here.

feature::Integer, if you are specifying a variable, pass the index of the variable here.

DynamicExpressions.EquationModule.Node — Method

Node(op::Int, l::Node)

Apply unary operator op (enumerating over the order given) to Node l

DynamicExpressions.EquationModule.Node — Method

Node(op::Int, l::Node, r::Node)

Apply binary operator op (enumerating over the order given) to Nodes l and r

DynamicExpressions.EquationModule.Node — Method

Node(var_string::String)

Create a variable node, using the format "x1" to mean feature 1

When you create an Options object, the operators passed are also re-defined for Node types. This allows you use, e.g., t=Node("x1") * 3f0 to create a tree, so long as * was specified as a binary operator.

When using these node constructors, types will automatically be promoted. You can convert the type of a node using convert:

Base.convert — Method

convert(::Type{Node{T1}}, n::Node{T2}) where {T1,T2}

Convert a Node{T2} to a Node{T1}. This will recursively convert all children nodes to Node{T1}, using convert(T1, tree.val) at constant nodes.

You can set a tree (in-place) with set_node!:

DynamicExpressions.EquationModule.set_node! — Method

set_node!(tree::Node{T}, new_tree::Node{T}) where {T}

Set every field of tree equal to the corresponding field of new_tree.

You can create a copy of a node with copy_node:

DynamicExpressions.EquationModule.copy_node — Method

copy_node(tree::Node; preserve_topology::Bool=false)

Copy a node, recursively copying all children nodes. This is more efficient than the built-in copy. With preserve_topology=true, this will also preserve linkage between a node and multiple parents, whereas without, this would create duplicate child node copies.